ACLU Review Template

Apply ACLU Analytics team review templates to your github repository to help standardize reviews and ensure quality and accuracy of our work

use_aclu_review_template(type = "analysis", repo = ".")

use_aclu_pr_template(type = "analysis", repo = ".")

use_aclu_issue_template(type = "analysis", repo = ".")

use_aclu_label_template(type = "analysis", repo = ".")

Arguments

type
Type of repository. Defaults to "analysis" which is the only fully built out template in base version

repo
Repository. Defaults to "." - base path. We use `repo` instead of `path` here to align with `git2r` standards.

Value

invisible. Provides cli based messages as completes

Examples

# apply all of the template items
use_aclu_review_template()

#> ✖ ./.github/pull_request_template.md already exists. Skipping file copy.
#> ✔ Pull request template applied to repository
#> ✖ ./.github/ISSUE_TEMPLATE/coding_practices.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/correctness_approach.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/data_accuracy.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/documentation.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/performance.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/readability.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/reproducibility.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/results_accuracy.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/results_completeness.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/suggestions.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/config.yml already exists. Skipping file copy.
#> ✔ Issue templates applied to repository
#> ✖ ./.github/LABEL_TEMPLATE/config.yml already exists. Skipping file copy.
#> ✔ Label templates added to .github folder

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> ℹ No labels need renaming
#> ✔ Adding missing labels: 'Data Accuracy :microscope:', 'Correctness of Approach :crystal_ball:', 'Results Accuracy :telescope:', 'Results Completeness :gem:', 'Reproducibility :ledger:', 'Coding Practices :art:', 'Readability :book:', 'Performance :hourglass:', 'Documentation :memo:', 'Enhancement Opportunities :school_satchel:', 'Alternatives to Consider :bulb:'
#> ℹ Label colours are up-to-date
#> ℹ Label descriptions are up-to-date
#> ✔ Label templates in .github folder applied to repository

# apply the PR template to the repository
use_aclu_pr_template()

#> ✖ ./.github/pull_request_template.md already exists. Skipping file copy.
#> ✔ Pull request template applied to repository

# apply issue templates to the repository
use_aclu_issue_template()

#> ✖ ./.github/ISSUE_TEMPLATE/coding_practices.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/correctness_approach.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/data_accuracy.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/documentation.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/performance.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/readability.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/reproducibility.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/results_accuracy.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/results_completeness.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/suggestions.md already exists. Skipping file copy.
#> ✖ ./.github/ISSUE_TEMPLATE/config.yml already exists. Skipping file copy.
#> ✔ Issue templates applied to repository

# apply labels to the repository
use_aclu_label_template()

#> ✖ ./.github/LABEL_TEMPLATE/config.yml already exists. Skipping file copy.
#> ✔ Label templates added to .github folder

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> Warning: Vectorizing 'git_emoji' elements may not preserve their attributes

#> ℹ No labels need renaming
#> ✔ Adding missing labels: 'Data Accuracy :microscope:', 'Correctness of Approach :crystal_ball:', 'Results Accuracy :telescope:', 'Results Completeness :gem:', 'Reproducibility :ledger:', 'Coding Practices :art:', 'Readability :book:', 'Performance :hourglass:', 'Documentation :memo:', 'Enhancement Opportunities :school_satchel:', 'Alternatives to Consider :bulb:'
#> ℹ Label colours are up-to-date
#> ℹ Label descriptions are up-to-date
#> ✔ Label templates in .github folder applied to repository